requirements.txt
streamlit==1.52.0
pandas==2.0.0
Pillow==9.5.0
qrcode==8.2
numpy==1.26.4
tensorflow==2.15.1
h5py==3.9.0
